train/requirements-train.txt
# Pilot training requirements (for vast.ai H100 instance)
# Compatible with PyTorch 2.4 + CUDA 12.1
# Install with: pip install -r requirements-train.txt

# Core ML
torch>=2.4.0
transformers>=4.45.0   # required for Qwen2.5 + SeamlessM4T-v2
peft>=0.13.0
accelerate>=1.0.0
datasets>=3.0.0
tokenizers>=0.20.0     # Qwen tokenizer

# Optional: bitsandbytes for 4-bit Qwen if memory becomes tight
# bitsandbytes>=0.43.0

# SeamlessM4T-v2 needs sentencepiece tokenizer
sentencepiece>=0.2.0
protobuf>=4.25.0

# Audio
librosa>=0.10.2
soundfile>=0.12.1
audioread>=3.0.0

# Evaluation
sacrebleu>=2.4.0
bert-score>=0.3.13
evaluate>=0.4.3
jiwer>=3.0.5

# Logging
wandb>=0.18.0
pyyaml>=6.0
tqdm>=4.66.0

# HF Hub
huggingface-hub>=0.25.0
